+++ solc/CommandLineParser.h
@@ -174,6 +174,7 @@ struct CommandLineOptions
 		std::vector<boost::filesystem::path> includePaths;
 		FileReader::FileSystemPathSet allowedDirectories;
 		bool ignoreMissingFiles = false;
+		bool noImportCallback = false;
 	} input;
 
 	struct